Коммит

+++ Ebay/Server/Data/Models/Purchase.cs
@@ -7,7 +7,7 @@ public class Purchase
     [Key]
     public DateTime Date { get; set; }
 
-    public int LotId { get; set; }
+    public long LotId { get; set; }
     public Lot Lot { get; set; } = null!;
 
     public double? Price { get; set; }
